Spotify player uses full playlist rather than individual play buttons
